Feature/as 3211 run liquibase against new schema (#27)
* AS-3211: Fix the order of instructions in README.md.

* AS-3211: Change database port to a placeholder in README.md.

* AS-3211: Enable allow bean definition overriding.

* AS-3211: Correct table names.

* AS-3211: Correct column names.

* AS-3211: Fix datatype mismatch issue.

* AS-3211: Correct column names.

* AS-3211: Fix datatype mismatch issue.

* AS-3211: Fix datatype mismatch issue.

* AS-3211: Correct table names.

+++ farms-api/farms-persistence/src/main/java/ca/bc/gov/farms/persistence/v1/dao/mybatis/ImportDaoImpl.java
@@ -31,7 +31,7 @@ public class ImportDaoImpl extends BaseDao implements ImportDao {
 
         try (CallableStatement callableStatement = this.conn
                 .prepareCall("call farms_webapp_pkg.insert_import_version(?, ?, ?, ?, ?, ?, ?, ?)")) {
-            callableStatement.registerOutParameter(1, Types.NUMERIC);
+            callableStatement.registerOutParameter(1, Types.BIGINT);
             callableStatement.setString(2, dto.getImportClassCode());
             callableStatement.setString(3, dto.getImportStateCode());
             callableStatement.setString(4, dto.getDescription());
@@ -40,7 +40,7 @@ public class ImportDaoImpl extends BaseDao implements ImportDao {
             callableStatement.setBytes(7, dto.getImportFile());
             callableStatement.setString(8, dto.getImportedByUser());
             callableStatement.execute();
-            dto.setImportVersionId(callableStatement.getBigDecimal(1).longValue());
+            dto.setImportVersionId(callableStatement.getLong(1));
         } catch (RuntimeException | SQLException e) {
             handleException(e);
         }
